Clean up codebase for production
A .env file is now required to connect to the database so credentials aren't stored in the codebase.
The database creation script now only grants select, insert, update and delete privileges. It
also requires the creation of 'movnetuser' before it's run.
Source maps are not generated anymore because they were causing problems for running JS.